Moving all mock component configs to mock_configs module
